fix: bug with score
This commit fixes a bug that would cause the AI player's score to start at 1 upon repeat games or when having lingered on the main menu for 5 seconds or more. This issue was caused by initialising lastFrameTime as 0.0 instead of Time.getTime() in the run() methods. This would cause deltaTime to be miscalculated when the Window thread starts; in the first iteration of run(), a huge deltaTime() was passed to ball.update(dt), causing the ball to move off the left edge of the screen, immediately incrementing the score as 1 for the AI player.